Intrusion_ Python Project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
facenet-pytorch==2.6.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pillow==10.2.0
propcache==0.3.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.1
six==1.17.0
supervision==0.26.1
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
twilio==9.7.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
